pacland driver and some fixes to driverH

+++ emulator/src/main/java/gr/codebb/arcadeflex/v037b16/drivers/galaga.java
@@ -340,7 +340,7 @@ public class galaga {
                         galaga_interrupt_3, 2
                 )
             },
-            (int) 60.606060, DEFAULT_60HZ_VBLANK_DURATION, /* frames per second, vblank duration */
+            60.606060, DEFAULT_60HZ_VBLANK_DURATION, /* frames per second, vblank duration */
             99, /* 99 CPU slices per frame - with 100, galagab2 hangs on coin insertion */
             galaga_init_machine,
             /* video hardware */
